backend_py/requirements.txt
# Core FastAPI dependencies
fastapi==0.103.1
starlette==0.27.0
anyio==3.7.1
sniffio==1.3.1
h11==0.14.0
typing_extensions==4.15.0

# Pydantic for data validation
pydantic==1.10.19
email-validator==2.3.0

# ASGI server for running the application
uvicorn==0.24.0

# Testing
pytest==8.4.2
httpx==0.24.1

# Environment management
python-dotenv==1.0.0

# Development tools (optional - can be removed for production)
# ruff @ file:///C:/Users/User/Downloads/ruff-0.14.1-py3-none-win_amd64.whl#sha256=59d599cdff9c7f925a017f6f2c256c908b094e55967f93f2821b1439928746a1
